Make 'enum NavigationType' be an enum class https://bugs.webkit.org/show_bug.cgi?id=144270
Reviewed by Alex Christensen.

Source/WebCore:

No new tests (No change in behavior).

I’ll be adding a new NavigationType for bug 144269 so it makes sense to make this
an enum class first.

* loader/FrameLoader.cpp:
(WebCore::FrameLoader::load):
(WebCore::FrameLoader::reload):
(WebCore::FrameLoader::loadDifferentDocumentItem):
* loader/FrameLoaderTypes.h:

* loader/NavigationAction.cpp:
(WebCore::navigationType):
(WebCore::NavigationAction::NavigationAction):

* loader/PolicyChecker.cpp:
(WebCore::PolicyChecker::checkNavigationPolicy):

* page/PerformanceNavigation.cpp:
(WebCore::PerformanceNavigation::type):

Source/WebKit/mac:

* WebCoreSupport/WebFrameLoaderClient.mm:
(WebFrameLoaderClient::actionDictionary):

Source/WebKit2:

* Shared/API/c/WKSharedAPICast.h:
(WebKit::toAPI):

* Shared/NavigationActionData.cpp:
(WebKit::NavigationActionData::NavigationActionData):

* UIProcess/API/Cocoa/WKNavigationAction.mm:
(toWKNavigationType):

* UIProcess/API/gtk/WebKitPrivate.cpp:
(toWebKitNavigationType):

* WebProcess/WebCoreSupport/WebFrameLoaderClient.cpp:
(WebKit::WebFrameLoaderClient::dispatchDecidePolicyForNavigationAction):

* WebProcess/WebPage/WebInspector.cpp:
(WebKit::WebInspector::openInNewTab):

Canonical link: https://commits.webkit.org/162269@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@183405 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebKit/mac/WebCoreSupport/WebFrameLoaderClient.mm
@@ -1566,7 +1566,7 @@ NSDictionary *WebFrameLoaderClient::actionDictionary(const NavigationAction& act
     NSURL *originalURL = action.url();
 
     NSMutableDictionary *result = [NSMutableDictionary dictionaryWithObjectsAndKeys:
-        [NSNumber numberWithInt:action.type()], WebActionNavigationTypeKey,
+        [NSNumber numberWithInt:static_cast<int>(action.type())], WebActionNavigationTypeKey,
         [NSNumber numberWithInt:modifierFlags], WebActionModifierFlagsKey,
         originalURL, WebActionOriginalURLKey,
         nil];
